IPG-Agent/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.3.5
